prototype/requirements.txt
chromadb>=0.5.0
openai>=1.50.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
markdown>=3.5.2
